pdcom-parser/pyproject.toml
[project]
name = "pdcom-parser"
version = "0.1.0"
description = "Swiss PDCom (Plan directeur communal) vector map extractor — PDF → PostGIS"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "shapely>=2.0.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "psycopg[binary]>=3.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "rtree>=1.0.0",
    "tenacity>=8.2.0",
    "rapidfuzz>=3.6.0",
]

[project.scripts]
pdcom = "pdcom_parser.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
